Move powers to separate modules
